requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0